+++ app/api-flask/README.md
@@ -1,1 +1,4 @@
-flask\Scripts\pip install -r requirements.txt
+#Create Virtual environment
+python3 -m venv venv
+#Install dependencies
+venv\Scripts\pip install -r requirements.txt
