+++ app/api-flask/clean_cache.sh
@@ -1,0 +1,1 @@
+find . -name '__pycache__' | xargs rm -rf
