+++ app/api-flask/dev-requires.txt
@@ -1,0 +1,2 @@
+-r run-requires.txt
+-r tests-requires.txt
