+++ app/api-flask/docker-compose.yaml
@@ -1,0 +1,22 @@
+version: '3'
+
+services:
+  web:
+    build:
+      context: .
+      dockerfile: Dockerfile.dev
+    ports:
+      - "3000:3000"
+      - "5000:5000"
+    environment:
+      - FLASK_ENV=production
+      - USER=goat
+      - HOST=db
+      - DBPORT=5432
+      - FLASKPORT=5000
+      - PASSWORD=earlmanigault
+      - DBNAME=goat
+    volumes:
+      - .:/usr/src/app
+    networks:
+      proxy:
