+++ app/api-flask/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+python_paths = /usr/src/app/src
+filterwarnings= default
+                ignore:.*is deprecated.*:Warning
+                error::DeprecationWarning:importlib.*
