+++ app/api-flask/requirements.txt
@@ -1,10 +1,1 @@
-psycopg2
-loguru
-postgis
-ptvsd
-Flask
-flask_restful
-flask_cors
-marshmallow
-webargs
-pyyaml
+-r run-requires.txt
