+++ app/api-flask/run-requires.txt
@@ -1,0 +1,10 @@
+psycopg2
+loguru
+postgis
+ptvsd
+Flask
+flask_restful
+flask_cors
+marshmallow
+webargs
+pyyaml
