+++ app/api-flask/tests-requires.txt
@@ -1,0 +1,6 @@
+pytest
+pytest-cov
+pytest-xdist
+pytest-pythonpath
+pytest-warnings
+codecov
